Create parent Classifier class
